Task2/bot/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.6.2.post1
asgiref==3.8.1
astroid==3.3.5
async-timeout==4.0.3
attrs==24.2.0
autopep8==2.3.1
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==4.2.0
bidict==0.23.1
bleach==6.2.0
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.0
Cerberus==1.3.5
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.18
click==8.1.7
coloredlogs==15.0.1
cryptography==43.0.3
cssselect==1.2.0
dataclasses-json==0.6.7
dateparser==1.2.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dicttoxml==1.7.16
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
docutils==0.21.2
durationpy==0.9
eventlet==0.37.0
exceptiongroup==1.2.2
fastapi==0.115.4
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geoip2==4.8.0
geopy==2.4.1
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.154.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
h3==4.1.2
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ip2geotools==0.1.6
IP2Location==8.10.4
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.5.0
kubernetes==31.0.0
langchain==0.3.10
langchain-anthropic==0.3.0
langchain-community==0.3.5
langchain-core==0.3.22
langchain-openai==0.2.12
langchain-text-splitters==0.3.2
langsmith==0.1.140
lazy-object-proxy==1.10.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
maxminddb==2.6.2
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nh3==0.2.18
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.0
openai==1.57.2
opentelemetry-api==1.28.0
opentelemetry-exporter-otlp-proto-common==1.28.0
opentelemetry-exporter-otlp-proto-grpc==1.28.0
opentelemetry-instrumentation==0.49b0
opentelemetry-instrumentation-asgi==0.49b0
opentelemetry-instrumentation-fastapi==0.49b0
opentelemetry-proto==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-semantic-conventions==0.49b0
opentelemetry-util-http==0.49b0
orjson==3.10.11
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pip-review==1.3.0
pkginfo==1.10.0
platformdirs==4.3.6
posthog==3.7.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.0
pylint==3.3.1
pymongo==4.13.2
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
ratelim==0.1.6
readme_renderer==44.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rsa==4.9
SecretStorage==3.3.3
selenium==4.26.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.35
starlette==0.41.2
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.8.0
timezonefinder==6.5.7
tokenizers==0.20.3
tomli==2.0.2
tomlkit==0.13.2
tqdm==4.66.6
trio==0.27.0
trio-websocket==0.11.1
twine==5.1.1
typed-ast==1.5.5
typer==0.12.5
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.2
wrapt==1.16.0
wsproto==1.2.0
yarl==1.17.1
zipp==3.20.2
